Commit extra 2: Añadimos metodo imprimir en el main

+++ GestionGit/src/gestionGit/Operaciones.java
@@ -15,7 +15,7 @@ public class Operaciones {
 		return a*b;
 	}
 	
-	public double dividir (int a,int b) {
+	public double dividir (double a,double b) {
 		if (b==0)return 0;
 		return a/b;
 	}
